# doc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "WeFile API",
    "version": "0.0.1"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/upload": {
      "post": {
        "description": "用户上传",
        "tags": [
          "File"
        ],
        "summary": "Upload",
        "operationId": "Upload",
        "parameters": [
          {
            "type": "string",
            "description": "采用 jwt 验证方式",
            "name": "Authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UploadResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/upload/try_fast": {
      "post": {
        "description": "尝试快速上传",
        "tags": [
          "File"
        ],
        "summary": "TryFastUpload",
        "operationId": "TryFastUpload",
        "parameters": [
          {
            "type": "string",
            "description": "采用 jwt 验证方式",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "file_hash": {
                  "description": "全量 Hash",
                  "type": "string",
                  "x-go-name": "FileHash"
                },
                "file_sampling_hash": {
                  "description": "采用抽样 Hash 的方式计算 Hash 快速上传",
                  "type": "string",
                  "x-go-name": "FileSamplingHash"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UploadResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/responses/ConflictError"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
...
  "responses": {
...
    "BadRequestResponse": {
      "description": "BadRequest",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
    "ConflictError": {
      "description": "Conflict Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "description": "Sever Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "NotFoundError": {
      "description": "NotFound Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "UnauthorizedResponse": {
      "description": "Unauthorized Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "UploadResponse": {
      "description": "Upload Response",
      "headers": {
        "file_id": {
          "type": "integer",
          "format": "int64",
          "description": "文件 id 用于下载"
        },
        "file_name": {
          "type": "string",
          "description": "文件名"
        },
        "file_size": {
          "type": "integer",
          "format": "int64",
          "description": "文件大小 单位 Byte"
        },
        "is_directory": {
          "type": "boolean",
          "description": "是否是目录"
        },
        "upload_at": {
          "type": "string",
          "format": "date-time",
          "description": "上传时间"
        }
      }
    },
...
  }
}